Aligned Vue and React examples.
