+++ examples/react/src/index.html
@@ -27,6 +27,6 @@
       To begin the development, run `yarn dev`.
       To create a production bundle, use `yarn build`.
     -->
-    <script src="./index.tsx"></script>
+    <script src="./main.tsx"></script>
   </body>
 </html>
